requirement.txt
python-telegram-bot==20.8
openai==1.3.7
google-generativeai==0.3.2
python-dotenv==1.0.0
psutil==5.9.5
pysqlcipher3==1.1